requirements.txt
fastapi
uvicorn
pandas
scikit-learn
nltk
python-multipart
python-dotenv
requests
sentence-transformers